refactor(storage): extract store auxiliary types
