+++ crates/storage/src/lib.rs
@@ -1,6 +1,10 @@
 mod api;
 pub mod backend;
+mod config;
 mod store;
+mod types;
+mod utils;
 
 pub use api::{ALL_TABLES, StorageBackend, StorageReadView, StorageWriteBatch, Table};
-pub use store::{ForkCheckpoints, GetForkchoiceStoreError, Store};
+pub use store::{GetForkchoiceStoreError, Store};
+pub use types::{ForkCheckpoints, GossipSignatureSnapshot};
